move Modify Action Maps to GameUiContext
