+++ SolastaUnfinishedBusiness/Models/BootContext.cs
@@ -19,22 +19,7 @@ internal static class BootContext
         ItemDefinitionVerification.Load();
         EffectFormVerification.Load();
 #endif
-
-        #region Modify Action Maps
-
-        var service = ServiceRepository.GetService<IInputService>();
-
-        //copy `GamepadSelector` action from `CharacterEdition` map into `ModalListBrowse` - needed for save by location to be able to scroll through save location selector
-        var map = service.InputActionAsset.FindActionMap("ModalListBrowse");
-        var action = map.AddAction("GamepadSelector");
-
-        var oldMap = service.InputActionAsset.FindActionMap("CharacterEdition").FindAction("GamepadSelector");
-        foreach (var oldMapBinding in oldMap.bindings)
-        {
-            action.AddBinding(oldMapBinding);
-        }
-
-        #endregion
+        GameUiContext.ModifyActionMaps();
 
         // STEP 0: Cache TA definitions for diagnostics and export
         DiagnosticsContext.CacheTaDefinitions();
